docs: fix stale fp32-default claims for opt state / LoRA adapter dtypes
The default ``state_dtype`` for ``AdamW`` / ``Muon`` and the default
``opt_state_dtype`` on ``TensorSpec`` are bf16 today (see
``flextrain/optim/adamw.py:73`` and ``flextrain/core/layer.py:132``),
but several comments and docs still claimed fp32. The actual code
behavior is unchanged -- only the documentation around it.

Found while reviewing memory-budget heuristics: the user-facing
``docs/working_set.md`` table claimed ``8 N`` (AdamW fp32) for the
opt-state row, which would have over-estimated memory usage by 2x
for anyone reading the table to size a run. Same drift in:

* ``flextrain/optim/adamw.py``: comment block before
  ``_ADAMW_STATE_SPEC_FP32`` claimed bf16 was the override; reversed
  to reflect bf16 default + when to override to fp32. Fixed the
  class docstring example to show the more useful override (fp32 for
  cold-start at scale) instead of the no-op (bf16 == default).
* ``flextrain/core/layer.py``: ``TensorSpec`` and ``ParamSpec``
  docstrings said "fp32 by default" for ``opt_state_dtype``. Now bf16.
* ``flextrain/api.py``: ``from_pretrained`` ``lora_targets`` docstring
  said the default LoRA adapter dtype is fp32 master/grad/opt-state
  (matching HF PEFT). The actual defaults are
  ``lora_adapter_*_dtype = torch.bfloat16``. Reframed -- bf16 is the
  memory-conservative default; fp32 is HF PEFT's at-scale convention
  and an opt-in override.
* ``docs/optimizers.md``: reversed default + override in the example.
* ``docs/working_set.md``: rewrote the heuristics table with the
  actual bf16 stack (4N AdamW / 2N Muon for opt state), with a note
  on the fp32 override doubling those.
* ``docs/internal/NOTES.md``: design-decision note still claimed
  ``opt_state_dtype`` defaults to float32. Now bfloat16.
* ``docs/internal/SESSION_NOTES.md``: Q-2 had the same bad parenthetical
  and was an open question that's already been resolved by the bf16
  default. Marked resolved.

No code or test changes -- defaults verified bf16 via direct assert
in adamw + TensorSpec.

+++ flextrain/optim/adamw.py
@@ -46,11 +46,10 @@ def _make_adamw_state_spec(dtype: torch.dtype) -> OptimizerStateSpec:
     )
 
 
-# Default: fp32 moments. This is the safer choice for numerical
-# stability in long runs. bf16 opt state is an available memory trade
-# (half the opt ring size) for large models where you'd otherwise need
-# host offload. User-selectable via the ``state_dtype`` constructor
-# arg.
+# bf16 moments are the default (cheap; halves opt-state memory). Pass
+# ``state_dtype=torch.float32`` for higher-precision moment updates --
+# recommended for cold-start pretraining at scale where the m / v
+# values can lose precision. Pre-built fp32 spec for that path.
 _ADAMW_STATE_SPEC_FP32 = _make_adamw_state_spec(torch.float32)
 
 
@@ -60,8 +59,8 @@ class AdamW(Optimizer):
     Usage
     -----
         opt = AdamW(hp=AdamWHyperparams(lr=3e-4))
-        # or bf16 opt state to halve memory:
-        opt = AdamW(hp=..., state_dtype=torch.bfloat16)
+        # or fp32 opt state for cold-start pretraining at scale:
+        opt = AdamW(hp=..., state_dtype=torch.float32)
         state = {}  # engine-allocated per (param, state_tensor) pair
         opt.step(param_spec, master, grads, state, step_num=step)
     """
